Add files via upload
This update has a main logic loop with a string, a for loop and some pseudo-code for the printing of the symbols. 
It also uses C's sizeof() function to find the length of the string. This will be needed as I learned that C does not have a For Each Loop function. 

An integer j is set to be the value of the letter in the string, depending on what the value of j is a series of if else statements will decide what the program will print.