Complete Version 1 With Picture
This update does not include any code changes but does have the worked string that when read prints a outline of Washington State. Potential for future additions could include a way to type in the picture and print the Encoded string. The challenge here would be finding a way to get around the Enter function when working in console. Possibly that each line is held within an array? The goal of this program was to have some more experience programming in C which I did.